📝 Improve changelog and upgrade guide (#450)
## Description

This PR improves the changelog and upgrade guide with regard to the
discontinuation of x86 macOS support. Support was initially dropped in
#421. See https://github.com/munich-quantum-toolkit/core/pull/1165/ for
more information.

Fixes #432

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] ~I have added appropriate tests that cover the new/changed
functionality.~
- [x] ~I have updated the documentation to reflect these changes.~
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.