👷 Fix `minimums` nox session (#468)
## Description

This PR fixes the `minimums` nox session by setting a lower bound for
Qiskit. The sessions is currently failing (including in the CI on
`main`) because MQT Predictor is not yet compatible with the latest
Qiskit version.

As a lock-file update was necessary, this PR also makes MQT Predictor
compatible with the latest Qiskit version.

## Checklist:

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] ~I have added appropriate tests that cover the new/changed
functionality.~
- [x] ~I have updated the documentation to reflect these changes.~
- [x] ~I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.~
- [x] ~I have added migration instructions to the upgrade guide (if
needed).~
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: burgholzer <burgholzer@me.com>